requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
bcrypt==4.0.1 
python-dotenv==1.1.0